refactor: - converted to pytest now - renamed scripts so clear which is for testing functions and which is for testing - created markdown for testing guide
